requirements.txt
Django>=3.2.0,<3.2.5
djangorestframework>=3.12.0,<3.12.4
flake8==3.9.2
pyzk>=0.5
requests==2.26.0
django-extensions==3.1.3
tenacity==8.0.1
django_environ==0.8.1